+++ lib/Entry.js
@@ -1,6 +1,4 @@
-// import fs from 'fs-extra'
 import path from 'path';
-import { fileURLToPath } from 'url';
 
 import postcss from 'postcss'
 import cssnano from 'cssnano'
