+++ lib/atrules/InlineComponentRule.js
@@ -42,32 +42,3 @@ export default class InlineComponentRule extends ComponentRule {
     })
   }
 }
-
-// export default class InlineComponentRule extends AtRule {
-//   constructor (atrule) {
-//     super({
-//       root: atrule,
-//       format: '<component>',
-//       args: [
-//         {
-//           name: 'component',
-//           types: ['word'],
-//           required: true
-//         }
-//       ]
-//     })
-//
-//     this.name = NGN.DATA.util.GUID()
-//     this.parentRule = this.#getParent(this.root)
-//   }
-//
-//   #getParent = (child, cb) => {
-//     let { parent } = child
-//
-//     if (parent.type !== 'rule') {
-//       return this.#getParent(parent)
-//     }
-//
-//     return parent
-//   }
-// }
