+++ lib/atrules/import/FileResource.js
@@ -5,7 +5,6 @@ import { CONFIG } from '../../../index.js'
 import Resource from './Resource.js'
 
 import FileUtils from '../../utilities/FileUtils.js'
-import QueueUtils from '../../utilities/QueueUtils.js'
 
 export default class FileResource extends Resource {
   #filepath
