+++ lib/atrules/set/Setting.js
@@ -36,25 +36,3 @@ export default class Setting {
     return !!typeset ? new Typeset(typeset) : null
   }
 }
-
-// export default class Setting {
-//   #root = null
-//
-//   constructor (atrule) {
-//     this.#root = atrule
-//     this.source = atrule.source
-//     this.bounds = atrule.bounds
-//     this.selector = atrule.selector
-//     this.margin = atrule.margin
-//     this.padding = atrule.padding
-//     this.typeset = atrule.typeset
-//   }
-//
-//   get root () {
-//     return this.#root
-//   }
-//
-//   replaceWith (nodes) {
-//     this.#root.replaceWith(nodes)
-//   }
-// }
