+++ lib/data/Config.js
@@ -2,10 +2,8 @@ import ErrorUtils from '../utilities/ErrorUtils.js'
 import FileUtils from '../utilities/FileUtils.js'
 
 import ConfigModel from './models/ConfigModel.js'
-import Constants from './Constants.js'
 import Defaults from './Defaults.js'
 import Functions from '../Functions.js'
-import TypographyUtils from '../utilities/TypographyUtils.js'
 
 import CoreModule from '../modules/core.js'
 import ComponentsModule from '../modules/components.js'
@@ -173,19 +171,6 @@ export default class Config {
     this.#model.load(defaults)
   }
 
-  // storeApplication (viewportName, application, cb) {
-  //   let viewport = this.#viewports.find({ name: viewportName })[0]
-  //
-  //   if (!viewport) {
-  //     return cb(ErrorUtils.createError({
-  //       message: `DEVELOPER ERROR: Viewport "${viewportName}" not found`
-  //     }))
-  //   }
-  //
-  //   viewport.applications.add(application)
-  //   cb()
-  // }
-
   #getBooleanValue = (prop, value) => {
     switch (typeof value) {
       case 'boolean': return value
@@ -306,64 +291,4 @@ export default class Config {
 
     return final
   }
-
-  // #storeGeneratedRanges = () => {
-  //
-  //
-  //   // this.generatedRanges = [
-  //   //   {
-  //   //     type: 'height',
-  //   //     name: 'min-height-and-over',
-  //   //     min: height.min
-  //   //   },
-  //   //
-  //   //   {
-  //   //     type: 'height',
-  //   //     name: 'min-height',
-  //   //     max: height.min
-  //   //   },
-  //   //
-  //   //   {
-  //   //     type: 'height',
-  //   //     name: 'over-min-height',
-  //   //     min: height.min + 1
-  //   //   },
-  //   //
-  //   //   {
-  //   //     type: 'width',
-  //   //     name: 'min-width-and-over',
-  //   //     min: width.min
-  //   //   },
-  //   //
-  //   //   {
-  //   //     type: 'width',
-  //   //     name: 'over-min-width',
-  //   //     min: width.min + 1
-  //   //   },
-  //   //
-  //   //   {
-  //   //     type: 'width',
-  //   //     name: 'under-max-width',
-  //   //     max: width.max - 1
-  //   //   },
-  //   //
-  //   //   {
-  //   //     type: 'width',
-  //   //     name: 'max-width-and-under',
-  //   //     max: width.max
-  //   //   },
-  //   //
-  //   //   {
-  //   //     type: 'width',
-  //   //     name: 'max-width-and-over',
-  //   //     min: width.max
-  //   //   },
-  //   //
-  //   //   {
-  //   //     type: 'width',
-  //   //     name: 'over-max-width',
-  //   //     min: width.max + 1
-  //   //   }
-  //   // ]
-  // }
 }
