+++ lib/data/models/TypographyConfigModel.js
@@ -1,6 +1,4 @@
-import Constants from '../Constants.js'
 import Defaults from '../Defaults.js'
-import TypographyUtils from '../../utilities/TypographyUtils.js'
 
 export default new NGN.DATA.Model({
   autoid: false,
