+++ lib/data/models/ViewportModel.js
@@ -1,5 +1,3 @@
-import Defaults from '../Defaults.js'
-
 function compare (value, comparator, condition) {
   return value
     ? comparator ? condition : true
