--- lib/data/stores/PartialStore.js
@@ -1,3 +1,0 @@
-export default class PartialManager {
-  
-}
