+++ lib/modules/components.js
@@ -1,4 +1,3 @@
-import parser from 'postcss-scss'
 import ComponentUtils from '../utilities/ComponentUtils.js'
 
 import anchor from '../components/anchor.js'
