--- lib/plugins/atrules/extend.js
@@ -1,12 +1,0 @@
-import postcss from 'postcss'
-
-export default postcss.plugin('chassis-atrules-extend', () => {
-  return (root, result) => new Promise((resolve, reject) => {
-    root.walkAtRules('extend', atrule => {
-      console.log('extend');
-      console.log(atrule.toString())
-    })
-
-    resolve(root)
-  })
-})
