+++ lib/plugins/handleErrors.js
