+++ lib/themes/default.js
@@ -47,9 +47,9 @@ const raw = parser.parse(`
       }
 
       button {
+        @set padding;
         background: lightgrey;
         border-radius: 3px;
-        @set padding;
       }
 
       table {
