+++ lib/utilities/ComponentUtils.js
@@ -1,8 +1,4 @@
 import parser from 'postcss-scss'
-// import nesting from 'postcss-nesting'
-// import CSSUtils from './CSSUtils.js'
-// import QueueUtils from './QueueUtils.js'
-// import SelectorUtils from './SelectorUtils.js'
 
 import ComponentRule from '../atrules/component/ComponentRule.js'
