--- lib/utilities/MathUtils.js
@@ -1,3 +1,0 @@
-export default class MathUtils {
-
-}
